+++ TestModule/TestModule.vcxproj
@@ -105,7 +105,7 @@
     <ClInclude Include="TestFunctions.h" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\op2ext.vcxproj">
+    <ProjectReference Include="..\src\op2ext.vcxproj">
       <Project>{6de3610c-c9ee-4fd3-bbd7-66382c16fde9}</Project>
     </ProjectReference>
   </ItemGroup>
