+++ src/GlobalDefines.cpp
