+++ src/StringConversion.cpp
